converts codegen scripts to TS
