+++ .gitignore
@@ -328,3 +328,5 @@ src/ocpp20/types.ts
 src/ocpp20/schemas.ts
 src/ocpp21/types.ts
 src/ocpp21/schemas.ts
+
+codegen/**/*.js
